// IhaleTakip(Radyoloji)/RadyolojiTakipUi/src/app/entity/model/tbl-ihale-islem.ts
export class TblIhaleIslem {
    ref_no: number;
    ihale_cari_kodu: string;
    ihale_cari_ad_unvan: string;
    ihale_hastane_cari_kodu: string;
    ihale_hastane_cari_ad_unvan: string;
    hasta_kabul_no: string;
    hasta_tc_no: string;
    hasta_adi: string;
    hasta_soyadi: string;
    cekim_tarihi: Date;
    cekim_saati: string;
    islem_kodu: string;
    islem_adi: string;
    islem_puani: number;
    cihaz_adi: string;
    islem_durumu: string;
    kayit_tarihi: Date;
    kayit_yapan: string;
    rapor_yazim_tarihi: Date;
    rapor_yazan_cari_kodu: string;
    rapor_yazan_cari_ad_unvan: string;
    rapor_onay_tarihi: Date;
    rapor_onay_cari_kodu: string;
    rapor_onay_cari_ad_unvan: string;
}